refactor: rename spinner component
