+++ providers/global-spinner.provider.tsx
@@ -2,13 +2,12 @@
 
 import { ReactNode } from 'react';
 
-import { SpinnerCustom } from '@/components/ui/custom-spinner';
+import { SpinnerCustom } from '@/components/ui/SpinnerCustom';
 import { useAuth } from '@/services/auth.service';
 
-export function AuthProvider({ children }: { children: ReactNode }) {
+export function GlobalSpinnerProvider({ children }: { children: ReactNode }) {
   const isAuthorizing = useAuth((state) => state.isAuthorizing);
 
-  // Global state control. Add your state check with ||
   const isLoading = () => isAuthorizing;
 
   if (isLoading())
